requirements.txt
dnspython
psycopg2-binary
ipwhois
pytz
python-dotenv
requests